pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "anomalous-field-recorder"
version = "0.4.0"
description = "Toolkit for logging and reporting anomalous electromagnetic field measurements"
authors = [{ name = "Field Research Team" }]
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
keywords = ["sensors", "analysis", "anomaly"]
dependencies = [
    "numpy>=1.26",
    "scipy>=1.11",
    "pandas>=2.1",
    "matplotlib>=3.8",
    "pydantic>=2.6",
    "scikit-learn>=1.3",
    "jinja2>=3.1",
    "fastapi>=0.110",
    "uvicorn>=0.23",
    "httpx>=0.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]
timescale = ["psycopg2-binary>=2.9"]
torch = ["torch>=2.1"]

[project.scripts]
afr = "afr.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.poetry.scripts]
afr = "afr.cli:main"
